add function for odometry

+++ src/main/include/angleMath.h
@@ -3,7 +3,7 @@
 #include <math.h>
 namespace am{
     // wrap input to between -pi and pi
-    void wrap(float &angle){
+    void wrap(double &angle){
         while (angle > M_PI){
             angle -= M_PI*2;
         }
@@ -13,7 +13,7 @@ namespace am{
     }
 
     // wrap input to between -180 and 180
-    void wrapDeg(float &angle) {
+    void wrapDeg(double &angle) {
         while (angle > 180){
             angle -= 360;
         }
